feat: wip new errors impl
